requirements.txt
fastapi
uvicorn
python-dotenv
pandas
tabulate
requests